requirements.txt
numpy==1.26.4
pillow==10.2.0
pandas==2.1.4
SciPy==1.11.4
